+++ Complexity_and_Scope/TodoAgility.Agile/TodoAgility.Agile.csproj
@@ -4,4 +4,10 @@
     <TargetFramework>netstandard2.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="CQRS\QueryHandlers" />
+    <Folder Include="Domain" />
+    <Folder Include="Persistence" />
+  </ItemGroup>
+
 </Project>
